adapt async logic
